chore: avoid memory leak in client
This commit fixes a memory leak in the client. The client was not reading the response body of the http request when closing the body, which could lead to a memory leak. This commit fixes this by reading the response body and discarding it.

This commit makes salus check pass.