Guarantee the order of transactions in the block endpoint (#489)
* feat: Guarantee the order of transactions in the block endpoint

This commit guarantees the order of transactions in the block endpoint. It does so by creating a map of fetched transactions and using that map to return the transactions in the original order they had in the block.

- Added a new struct `FetchedTransaction` to represent fetched transactions
- Updated `fetchChannelTransactions` to use the new `FetchedTransaction` struct
- Updated `UnsafeTransactions` to use the new `FetchedTransaction` struct and create a map of fetched transactions
- Modified the return statement in `UnsafeTransactions` to use the fetched transactions from the map

* chore: add timeout to mock server

This commit adds a timeout to the mock server to prevent it from hanging indefinitely. This is mainly to make salus security scans happy.

* chore: avoid memory leak in client

This commit fixes a memory leak in the client. The client was not reading the response body of the http request when closing the body, which could lead to a memory leak. This commit fixes this by reading the response body and discarding it.

This commit makes salus check pass.

* chore: make check-gen happy

* trigger github workflow